BreweryService.kt = Has been set up so that we can save our data to the SQLite database. We will be saving a breweries - name, street, city, state, phone, and website_url and each row will be assigned an ID that will increment.
The next step is to figure out - in which way are we going to save our favorite brewery? Are we going to click on the brewery and present a dialog presenting a save option? Then once we figure the how, I can implement the code to actually save the data in the right spot.

Sidenote- https://api.openbrewerydb.org/v1/ - when put into a search bar, this will not load. However, https://api.openbrewerydb.org loads which I believe could be the true base url.

So I changed the base to - https://api.openbrewerydb.org
and then in the GET annotation added -  @GET("/v1/breweries")

This works just the same not if a big deal but I made the changes in-case you guys think it is a good idea