************************************** X- The build files have been updated so that you can run the Cloud FireStore      (However, you still need to create your own SHA-1 Key I believe as Jesse Stated)
X- The Cloud FireStore is called - Brewery Buddy

X-The UI design is just for testing purposes (we need to discuss how we want to implement the CRUD features)

X- There are three buttons - the code I added allows us to add documents/breweries to the cloud, delete documents/breweries from the cloud, and in a dialog box we can view all of our breweries in a list.

X- I still need to finish some of the logging information.

X- I haven't implemented the Authentication feature and I believe this should be the next call to action.

X- I didn't implement an update brewery feature as I don't believe it makes sense.